import modal from component
